+++ Assets/Scripts/Generators/StringGenerator.cs
@@ -41,6 +41,7 @@ namespace Ventura.Generators
 
         //public static FileStringGenerator Sites = new FileStringGenerator("names_sites_international");
         public static FileStringGenerator Sites = new FileStringGenerator("names_sites_italia");
+        public static FileStringGenerator FirstNames = new FileStringGenerator("names_people_toscana");
 
 
         protected FileStringGenerator(string sourceFile): this(new string[] { sourceFile }) { }