+++ Assets/Scripts/Unity/Behaviours/InventoryItemManager.cs
@@ -9,7 +9,7 @@ using Ventura.Unity.Graphics;
 namespace Ventura.Unity.Behaviours
 {
 
-    public class InventoryItemHandler : MonoBehaviour, IPointerEnterHandler, IPointerExitHandler
+    public class InventoryItemManager : MonoBehaviour, IPointerEnterHandler, IPointerExitHandler
     {
         public GameItem gameItem;
 
@@ -21,7 +21,7 @@ namespace Ventura.Unity.Behaviours
         void Start()
         {
             thumbnail.sprite = SpriteCache.Instance.GetSprite(gameItem);
-            labelText.text = gameItem.Name;
+            labelText.text = gameItem.Label;
 
             labelPanel.SetActive(false);
         }
@@ -29,7 +29,7 @@ namespace Ventura.Unity.Behaviours
 
         public void OnButtonClick()
         {
-            //Debug.Log($"InventoryItemHandler.OnButtonClick; gameItem.Name: {gameItem.Name}");
+            //Debug.Log($"InventoryItemManager.OnButtonClick; gameItem.Name: {gameItem.Name}");
 
             var orch = Orchestrator.Instance;
             var newAction = new UseAction(orch, orch.Player, gameItem);
@@ -39,11 +39,13 @@ namespace Ventura.Unity.Behaviours
 
         public void OnPointerEnter(PointerEventData eventData)
         {
+            Debug.Log($"InventoryItemManager.OnPointerEnter; gameItem.Name: {gameItem.Name}");
             labelPanel.SetActive(true);
         }
 
         public void OnPointerExit(PointerEventData eventData)
         {
+            Debug.Log($"InventoryItemManager.OnPointerExit; gameItem.Name: {gameItem.Name}");
             labelPanel.SetActive(false);
         }